refactor: Footer와 Header의 위치를 app하위로 수정
